requirements/local.txt
# Local development dependencies go here
-r base.txt
coverage==3.7.1
django-discover-runner==1.0
django-debug-toolbar==1.2.1
ipython==2.1.0
Sphinx==1.2.2
